# backend/app/api/v1/purchase_materials.py
from typing import Annotated

from fastapi import APIRouter, Depends, Query, status

from app.core.permissions import CurrentUser, DbSession, PurchaseWriter, require_roles
from app.domain.enums import Role
from app.models import User
from app.schemas import (
    LinkStockMaterialRequest,
    Page,
    PurchaseMaterialCreate,
    PurchaseMaterialRead,
    PurchaseMaterialUpdate,
)
from app.services import material_service
from app.services.common import validate_version

router = APIRouter(prefix="/purchase-materials", tags=["申购物资"])
PageNo = Annotated[int, Query(ge=1)]
PageSize = Annotated[int, Query(ge=1, le=200)]
LinkWriter = Annotated[
    User,
    Depends(require_roles(Role.SUPER_ADMIN, Role.WAREHOUSE_ADMIN, Role.PURCHASE_ADMIN)),
]


@router.get("", response_model=Page[PurchaseMaterialRead])
async def list_materials(
    session: DbSession,
    user: CurrentUser,
    page: PageNo = 1,
    page_size: PageSize = 20,
    keyword: str | None = None,
    enabled: bool | None = None,
    coded: bool | None = None,
) -> Page[PurchaseMaterialRead]:
    items, total = await material_service.search_purchase_materials(
        session, keyword=keyword, enabled=enabled, coded=coded, page=page, page_size=page_size
    )
    return Page(
        items=[await material_service.purchase_read(session, item) for item in items],
        page=page,
        page_size=page_size,
        total=total,
    )


@router.post("", response_model=PurchaseMaterialRead, status_code=status.HTTP_201_CREATED)
async def create_material(
    data: PurchaseMaterialCreate, session: DbSession, user: PurchaseWriter
) -> PurchaseMaterialRead:
    item = await material_service.create_purchase_material(session, data, user.id)
    return await material_service.purchase_read(session, item)


@router.get("/{material_id}", response_model=PurchaseMaterialRead)
async def material_detail(
    material_id: int, session: DbSession, user: CurrentUser
) -> PurchaseMaterialRead:
    return await material_service.purchase_read(
        session, await material_service.get_purchase_material(session, material_id)
    )


@router.patch("/{material_id}", response_model=PurchaseMaterialRead)
async def update_material(
    material_id: int,
    data: PurchaseMaterialUpdate,
    session: DbSession,
    user: PurchaseWriter,
) -> PurchaseMaterialRead:
    item = await material_service.get_purchase_material(session, material_id)
    item = await material_service.update_purchase_material(session, item, data, user.id)
    return await material_service.purchase_read(session, item)


@router.post("/{material_id}/link-stock-material", response_model=PurchaseMaterialRead)
async def link_stock_material(
    material_id: int,
    data: LinkStockMaterialRequest,
    session: DbSession,
    user: LinkWriter,
) -> PurchaseMaterialRead:
    item = await material_service.get_purchase_material(session, material_id)
    validate_version(data.version, item.version)
    stock = await material_service._validate_stock_link(session, data.stock_material_id)
    item.stock_material_id = data.stock_material_id
    item.stock_material = stock
    item.updated_by = user.id
    item.version += 1
    await session.flush()
    return await material_service.purchase_read(session, item)