# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 电气车间备件管理系统
  version: 1.0.0
paths:
...
  /api/v1/purchase-materials:
    get:
      tags:
      - 申购物资
      summary: List Materials
      operationId: list_materials_api_v1_purchase_materials_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: enabled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      - name: coded
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Coded
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_PurchaseMaterialRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - 申购物资
      summary: Create Material
      operationId: create_material_api_v1_purchase_materials_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseMaterialCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseMaterialRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-materials/{material_id}:
    get:
      tags:
      - 申购物资
      summary: Material Detail
      operationId: material_detail_api_v1_purchase_materials__material_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: material_id
        in: path
        required: true
        schema:
          type: integer
          title: Material Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseMaterialRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - 申购物资
      summary: Update Material
      operationId: update_material_api_v1_purchase_materials__material_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: material_id
        in: path
        required: true
        schema:
          type: integer
          title: Material Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseMaterialUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseMaterialRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-materials/{material_id}/link-stock-material:
    post:
      tags:
      - 申购物资
      summary: Link Stock Material
      operationId: link_stock_material_api_v1_purchase_materials__material_id__link_stock_material_post
      security:
      - HTTPBearer: []
      parameters:
      - name: material_id
        in: path
        required: true
        schema:
          type: integer
          title: Material Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkStockMaterialRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseMaterialRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ApiError:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - code
      - message
      - request_id
      title: ApiError
...
    CodeState:
      type: string
      enum:
      - UNCODED
      - CODED
      title: CodeState
...
    FileObjectRead:
      properties:
        id:
          type: integer
          title: Id
        original_name:
          type: string
          title: Original Name
        url:
          type: string
          title: Url
        mime_type:
          type: string
          const: image/png
          title: Mime Type
          default: image/png
        size_bytes:
          type: integer
          title: Size Bytes
        width:
          type: integer
          title: Width
        height:
          type: integer
          title: Height
      type: object
      required:
      - id
      - original_name
      - url
      - size_bytes
      - width
      - height
      title: FileObjectRead
...
    LinkStockMaterialRequest:
      properties:
        stock_material_id:
          type: integer
          title: Stock Material Id
        version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Version
      additionalProperties: false
      type: object
      required:
      - stock_material_id
      title: LinkStockMaterialRequest
...
    Page_PurchaseMaterialRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PurchaseMaterialRead'
          type: array
          title: Items
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - page
      - page_size
      - total
      title: Page[PurchaseMaterialRead]
...
    PurchaseMaterialCreate:
      properties:
        material_code:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Material Code
        name:
          type: string
          maxLength: 128
          minLength: 1
          title: Name
        model_spec:
          type: string
          maxLength: 255
          minLength: 1
          title: Model Spec
        unit_id:
          type: integer
          title: Unit Id
        remark:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Remark
        stock_material_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Stock Material Id
        image_ids:
          items:
            type: integer
          type: array
          maxItems: 9
          title: Image Ids
      additionalProperties: false
      type: object
      required:
      - name
      - model_spec
      - unit_id
      title: PurchaseMaterialCreate
    PurchaseMaterialRead:
      properties:
        id:
          type: integer
          title: Id
        material_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Material Code
        name:
          type: string
          title: Name
        model_spec:
          type: string
          title: Model Spec
        unit_id:
          type: integer
          title: Unit Id
        unit_name:
          type: string
          title: Unit Name
        remark:
          anyOf:
          - type: string
          - type: 'null'
          title: Remark
        stock_material_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Stock Material Id
        stock_material_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Stock Material Name
        code_state:
          $ref: '#/components/schemas/CodeState'
        enabled:
          type: boolean
          title: Enabled
        images:
          items:
            $ref: '#/components/schemas/FileObjectRead'
          type: array
          title: Images
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        version:
          type: integer
          title: Version
      type: object
      required:
      - id
      - name
      - model_spec
      - unit_id
      - unit_name
      - code_state
      - enabled
      - images
      - created_at
      - updated_at
      - version
      title: PurchaseMaterialRead
    PurchaseMaterialUpdate:
      properties:
        material_code:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Material Code
        name:
          type: string
          maxLength: 128
          minLength: 1
          title: Name
        model_spec:
          type: string
          maxLength: 255
          minLength: 1
          title: Model Spec
        unit_id:
          type: integer
          title: Unit Id
        remark:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Remark
        stock_material_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Stock Material Id
        image_ids:
          items:
            type: integer
          type: array
          maxItems: 9
          title: Image Ids
        version:
          type: integer
          title: Version
      additionalProperties: false
      type: object
      required:
      - name
      - model_spec
      - unit_id
      - version
      title: PurchaseMaterialUpdate
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer